python updated to work with match. And sessionManagerUpdated
