+++ transcendence/Dockerfile
@@ -1,6 +1,6 @@
 # syntax=docker/dockerfile:1
 
-FROM python:3.9.18-bookworm
+FROM python:3.13-rc-bookworm
 
 RUN apt update && apt install -y locales apache2 apache2-dev postgresql-client
